pyproject.toml
[project]
name = "job-offer-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.42.80",
    "data-eng",
    "dbt-core>=1.10.20",
    "minio>=7.2.20",
    "polars>=1.36.1",
    "pre-commit>=4.3.0",
    "prefect>=3.4.25",
    "prefect-aws>=0.6.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.8",
    "urllib3<2",
]

[tool.uv.workspace]
members = [
    "data-eng",
]

[tool.uv.sources]
data-eng = { workspace = true }
